Move DiscoverLinks into Links
